refactor and fix refresh token
