+++ client/packages/common/src/authentication/api/hooks/index.ts
@@ -1,3 +1,6 @@
 export * from './useAuthApi';
 export * from './useGetRefreshToken';
 export * from './useUserDetails';
+export * from './useRefreshingAuth';
+export * from './useLogin';
+export * from './useGetUserPermissions';
