+++ client/packages/common/src/authentication/api/hooks/useUserDetails.ts
@@ -2,7 +2,8 @@ import { UserNode } from '@common/types';
 import { useMutation, useQuery } from 'react-query';
 import { useAuthApi } from './useAuthApi';
 
-export const useUserDetails = () => {
+export const useGetUserDetails = () => {
+  console.info(`************ GET USER DETAILS ************ `);
   const api = useAuthApi();
   return useMutation<
     Partial<UserNode> | undefined,
@@ -12,10 +13,10 @@ export const useUserDetails = () => {
   >(api.get.me);
 };
 
-export const useUserStores = (token: string) => {
+export const useUserDetails = (token: string) => {
+  console.log('---use---user---details---');
   const api = useAuthApi();
-  return useQuery(api.keys.me(token), api.get.stores(), {
-    cacheTime: 0,
+  return useQuery(api.keys.me(token), () => api.get.me(token), {
     enabled: !!token,
   });
 };
